Arreglo inicio sesion cliente por cambio de tipo dato
por el cambio de tipo de dato en identificación y en teléfono era necesario hacer cambios menores en el método obtenerClientesDB() del modelo cliente

+++ src/java/com/mallbit/cliente/ModeloCliente.java
@@ -34,8 +34,8 @@ public class ModeloCliente {
 	            String nombre = resultSet.getString("Nombre");
 	            String apellido = resultSet.getString("Apellido");
 	            String correo = resultSet.getString("Correo");
-                    int identificacion = resultSet.getInt("identificacion");
-                    int telefono = resultSet.getInt("telefono");
+                    long identificacion = resultSet.getLong("identificacion");
+                    long telefono = resultSet.getLong("telefono");
 	            String usuario = resultSet.getString("Usuario");
 	            String contraseña = resultSet.getString("Password");
 	            Date fechaNacimiento = resultSet.getDate("FechaNacimiento");
